raspberry_pi/gps/requirements.txt
pynmea2
pyserial